backend/requirements.txt
django>=4.2,<5.0
djangorestframework
gunicorn
psycopg2-binary>=2.9.3
dj_database_url>=0.5.0
django-cors-headers
djangorestframework-simplejwt
watchfiles
Pillow
requests>=2.25.0